# submition important files/swagger.yaml
openapi: 3.0.1
info:
  title: Sensor Data Processing API
  description: API for uploading and processing sensor data
  contact:
    name: Sensor Team
    email: support@sensor.com
  version: '1.0'
servers:
  - url: http://localhost:8081
    description: Generated server url

# GLOBAL SECURITY
security:
  - bearerAuth: []

tags:
  - name: Sensor Data Processing
    description: APIs for uploading, processing, and querying sensor data

paths:
  /api/v1/upload:
    post:
      tags:
        - Sensor Data Processing
      summary: Upload Sensor Data CSV
      description: Upload a CSV file containing sensor data for asynchronous processing.
      operationId: uploadData
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with timestamp_ms, device_id, channel, value
      responses:
        '202':
          description: File accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                success: true
                message: File uploaded successfully
                data:
                  uploadId: abc-123
                  filename: sensor_data.csv
                  fileSize: 1024
        '400':
          description: Bad request - Empty file or invalid file type
          content:
            application/json:
              example:
                success: false
                message: Only CSV files are allowed
                errorCode: INVALID_FILE_TYPE
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUploadResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUploadResponse'
      security:
        - bearerAuth: []


  /api/v1/auth/newUser:
    post:
      tags:
        - auth-controller
      summary: Register a new user
      operationId: addNewUser
      requestBody:
        description: New user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      responses:
        '200':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'

  /api/v1/auth/authenticate:
    post:
      tags:
        - auth-controller
      summary: Authenticate user and return JWT token
      operationId: authenticateAndGetToken
      requestBody:
        description: Username and password
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                type: string
        '403':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: string

  /api/v1/status:
    get:
      tags:
        - Sensor Data Processing
      summary: Get System Status
      operationId: getStatus
      responses:
        '200':
          description: Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStatusResponse'
      security:
        - bearerAuth: []

  /api/v1/results/{uploadId}:
    get:
      tags:
        - Sensor Data Processing
      summary: Get Processing Results
      operationId: getResults
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: query
          required: false
          schema:
            type: string
        - name: channel
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Results returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseResultResponse'
      security:
        - bearerAuth: []

  /api/v1/results/{uploadId}/summary:
    get:
      tags:
        - Sensor Data Processing
      summary: Get Results Summary
      operationId: getResultsSummary
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Summary returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
      security:
        - bearerAuth: []

  /api/v1/jobs:
    get:
      tags:
        - Sensor Data Processing
      summary: List All Jobs
      operationId: listJobs
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Jobs returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListMapStringObject'
      security:
        - bearerAuth: []

  /api/v1/jobs/{uploadId}:
    get:
      tags:
        - Sensor Data Processing
      summary: Get Job Details
      operationId: getJobInfo
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job info returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
      security:
        - bearerAuth: []

    delete:
      tags:
        - Sensor Data Processing
      summary: Delete Job
      operationId: deleteJob
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
      security:
        - bearerAuth: []

  /api/v1/health:
    get:
      tags:
        - Sensor Data Processing
      summary: Health Check
      operationId: health
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
      security:
        - bearerAuth: []

components:
  schemas:
    UploadResponse:
      type: object
      properties:
        uploadId:
          type: string
        message:
          type: string
        filename:
          type: string
        fileSize:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
        statusUrl:
          type: string

    ApiResponseUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UploadResponse'
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          type: string
        name:
          type: string
        password:
          type: string

    AuthenticationRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    StatusResponse:
      type: object
      properties:
        totalSamplesReceived:
          type: integer
          format: int64
        totalSamplesProcessed:
          type: integer
          format: int64
        totalInvalidSamples:
          type: integer
          format: int64
        pendingJobs:
          type: integer
          format: int64
        totalJobs:
          type: integer
          format: int64

    ApiResponseStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/StatusResponse'
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

    ResultResponse:
      type: object
      properties:
        uploadId:
          type: string
        status:
          type: string
        acceptedCount:
          type: integer
          format: int64
        rejectedCount:
          type: integer
          format: int64
        errorMessage:
          type: string
        statistics:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time

    ApiResponseResultResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/ResultResponse'
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

    ApiResponseMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

    ApiResponseListMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

    ApiResponseString:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: string
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time

  # FIXED securitySchemes
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT